requirements.txt
fastapi>=0.115.0
fastmcp>=3.0.0
uvicorn[standard]>=0.30.0
curl-cffi>=0.7.0
pydantic>=2.0.0
python-dotenv>=1.0.0
slowapi>=0.1.9
prometheus-fastapi-instrumentator>=7.0.0
torch>=2.0.0
torchvision>=0.15.0
Pillow>=10.0.0
